chore: create transfer index
